retail_insights_assistant/requirements.txt
streamlit
pandas
duckdb
python-dotenv
langgraph
langchain-core
langchain-google-genai
matplotlib
